Refactor: API 경로 구조 최적화 및 컨트롤러 분리
- 기존 복잡한 경로 구조를 RESTful한 단수형으로 간소화 (/api/calendar/todo/todo-types -> /api/todo/type 등)
- 컨트롤러 분리를 통한 코드 가독성 및 관리 효율성 개선
- 변경된 API 엔드포인트에 맞춰 프론트엔드 호출 코드 일괄 수정
