+++ backend/src/main/java/com/plands/backend/controller/CalendarController.java
@@ -23,7 +23,6 @@ import java.util.List;
 public class CalendarController {
 
     private final MemberService memberService;
-    private final CalendarService calendarService;
     private final TodoService todoService;
 
     // UserDetails에서 memberId를 안전하게 꺼내오는 든든한 헬퍼 메서드!
@@ -44,51 +43,7 @@ public class CalendarController {
                 .getMemberId();
     }
 
-    // 프론트(FullCalendar)가 요청하는 기간(startDate, endDate)을 파라미터로 직접 바인딩함
-    @GetMapping("/todo")
-    public ResponseEntity<List<CalendarResponseDto>> getTodoCalendarList(@AuthenticationPrincipal UserDetails userDetails, @RequestParam("startDate") String startDate, @RequestParam("endDate") String endDate) {
-
-        log.info("====== 달력 목록 조회 컨트롤러 진입 ======");
-
-        Long memberId = getAuthenticatedMemberId(userDetails);
-
-        log.debug("🔍 DB에서 조회된 진짜 회원 번호(memberId) -> {}", memberId);
-        log.debug("요청 파라미터 -> startDate: {}, endDate: {}", startDate, endDate);
-
-        // 서비스 레이어를 호출하여 한 달 치 데이터 가득 담긴 상자 더미 수령
-        List<CalendarResponseDto> todoList = calendarService.getCalendarList(memberId, startDate, endDate);
-
-        // 상태 코드 200(OK)과 함께 프론트엔드로 응답 전송
-        return ResponseEntity.ok(todoList);
-    }
-
-    // 새 할 일 등록 API
-    @PostMapping("/todo")
-    public ResponseEntity<String> createTodo(@AuthenticationPrincipal UserDetails userDetails, @RequestBody TodoRequestDto todoRequestDto) { // RequestBody는 http body 내의 json 속 데이터를 dto에 매핑
-
-        log.info("====== 할 일 등록 컨트롤러 진입 ======");
-        log.debug("프론트에서 넘어온 데이터: {}", todoRequestDto.toString());
-
-        Long memberId = getAuthenticatedMemberId(userDetails);
-        todoRequestDto.setMemberId(memberId);
-
-        // 💡 @RequestBody가 프론트에서 쏜 JSON 데이터를 자바 DTO 객체(참조변수 주소값)로 찰떡같이 변환해줘!
-        todoService.createTodo(todoRequestDto);
-
-        return ResponseEntity.ok("할 일이 성공적으로 등록되었습니다.");
-    }
-
-    // 할 일 종류 목록 전체 조회 API
-    @GetMapping("/todo-types")
-    public ResponseEntity<List<TodoTypeResponseDto>> getTodoTypes() {
-        log.info("====== 할 일 종류 조회 컨트롤러 진입 ======");
-
-        List<TodoTypeResponseDto> list = todoService.getTodoTypeList();
-
-        return ResponseEntity.ok(list);
-    }
-
-    @GetMapping("/member-plants")
+    @GetMapping("/member-plant")
     public ResponseEntity<List<MemberPlantResponseDto>> getMemberPlants(@AuthenticationPrincipal UserDetails userDetails) {
         log.info("====== 회원 식물 목록 조회 컨트롤러 진입 ======");
 
@@ -97,28 +52,5 @@ public class CalendarController {
         List<MemberPlantResponseDto> list = todoService.getMemberPlantList(memberId);
 
         return ResponseEntity.ok(list);
-    }
-
-    // 할 일 삭제 API (Soft Delete)
-    @DeleteMapping("/todo/{todoId}")
-    public ResponseEntity<String> deleteTodo(@PathVariable Long todoId) {
-        log.info("====== 할 일 삭제 컨트롤러 진입 ======");
-        log.debug("프론트에서 넘어온 삭제 대상 ID: {}", todoId);
-
-        todoService.deleteTodo(todoId);
-
-        return ResponseEntity.ok("할 일이 성공적으로 삭제되었습니다.");
-    }
-
-    // 할 일 수정 API
-    @PutMapping("/todo/{todoId}")
-    public ResponseEntity<String> updateTodo(@PathVariable Long todoId, @RequestBody TodoRequestDto todoRequestDto) {
-
-        log.info("====== 할 일 수정 컨트롤러 진입 ======");
-        log.debug("수정할 할 일 ID: {}", todoId);
-
-        todoService.updateTodo(todoId, todoRequestDto);
-
-        return ResponseEntity.ok("할 일이 성공적으로 수정되었습니다.");
     }
 }
